Fixed issue with data types. Integers stored as int rather than ASCII chars.

+++ Cart.h
@@ -6,14 +6,14 @@
 class Cart
 {
 private:
-	uint8_t size;
-	uint8_t items;
+	int size;
+	int items;
 
-	std::list<std::pair<uint8_t, Item>> contents; // List of pairs. First value of pair specifies item quantity
+	std::list<std::pair<int, Item>> contents; // List of pairs. First value of pair specifies item quantity
 
 public:
 	Cart();
-	void AddItem(uint8_t quantityIn, Item itemIn);
+	void AddItem(int quantityIn, Item itemIn);
 	void ReturnAllItems();
 	//~Cart();
 };
@@ -27,17 +27,17 @@ Cart::Cart()
 // Using 'quantity to stack multiple items in cart
 // Need to increment quantity
 // How can we search for items in the cart?
-void Cart::AddItem(uint8_t quantityIn, Item itemIn)
+void Cart::AddItem(int quantityIn, Item itemIn)
 {
 	// Using std::find_if algorithm with overloaded operator == for Item class
 	// Checks to see if the second property of the pair matches the item being added; disregards quantity at this point
-	auto it = std::find_if(contents.begin(), contents.end(), [&itemIn](const std::pair<uint8_t, Item>& p) { return p.second == itemIn; });
+	auto it = std::find_if(contents.begin(), contents.end(), [&itemIn](const std::pair<int, Item>& p) { return p.second == itemIn; });
 
 	// If iterator points to the end of the list then the item doesn't exist, add to end of list
 	// Else, increment value where iterator points
 	if (it == contents.end())
 	{
-		std::pair<size_t, Item> newItem(quantityIn, itemIn);
+		std::pair<int, Item> newItem(quantityIn, itemIn);
 		contents.push_back(newItem);
 		std::cout << std::to_string(quantityIn) << " " << itemIn.ReturnName() << " added to cart." << "\n";
 	}
